chore(agent-core): move tool services type
