+++ .changeset/tool-support-services.md
@@ -1,0 +1,6 @@
+---
+"@moonshot-ai/agent-core": patch
+"@moonshot-ai/kimi-code": patch
+---
+
+Relocate shared tool service typing to the tool support layer.
