--- packages/agent-core/src/runtime-types.ts
@@ -1,6 +1,0 @@
-import type { UrlFetcher, WebSearchProvider } from './tools/builtin';
-
-export interface ToolServices {
-  readonly urlFetcher?: UrlFetcher | undefined;
-  readonly webSearcher?: WebSearchProvider | undefined;
-}
