+++ packages/agent-core/src/tools/support/services.ts
@@ -1,0 +1,6 @@
+import type { UrlFetcher, WebSearchProvider } from '../builtin';
+
+export interface ToolServices {
+  readonly urlFetcher?: UrlFetcher;
+  readonly webSearcher?: WebSearchProvider;
+}
